+++ hummingbot/client/commands/__init__.py
@@ -1,8 +1,12 @@
+from .bounty_command import BountyCommand
 from .help_command import HelpCommand
+from .history_command import HistoryCommand
 from .start_command import StartCommand
 
 
 __all__ = [
+    BountyCommand,
     HelpCommand,
+    HistoryCommand,
     StartCommand,
 ]
