+++ hummingbot/client/commands/export_trades_command.py
@@ -1,0 +1,7 @@
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from hummingbot.client.hummingbot_application import HummingbotApplication
+
+
+class HelpCommand:
