+++ hummingbot/client/commands/get_balance_command.py
@@ -1,0 +1,8 @@
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from hummingbot.client.hummingbot_application import HummingbotApplication
+
+
+class GetBalanceCommand:
+    pass
